--- NorthstarDLL/context.h
@@ -1,11 +1,0 @@
-#pragma once
-
-enum class ScriptContext : int
-{
-	SERVER,
-	CLIENT,
-	UI,
-	NONE
-};
-
-const char* GetContextName(ScriptContext context);
