+++ NorthstarDLL/logging.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "context.h"
 
 void CreateLogFiles();
 void InitialiseLogging();
